# web-api/docs/swagger.yaml
basePath: '{{.BasePath}}'
definitions:
...
  model.AccountPwd:
    properties:
      confirm_pwd:
        type: string
      id:
        type: string
      new_pwd:
        type: string
      old_pwd:
        type: string
    type: object
  model.AccountUpdate:
    properties:
      address:
        type: string
      contact:
        type: string
      email:
        type: string
      id:
        type: string
      login_id:
        type: string
      nick_name:
        type: string
      phone:
        type: string
      remark:
        type: string
      type_id:
        type: string
      username:
        type: string
    type: object
  model.CreateAccount:
    properties:
      address:
        type: string
      c_time:
        type: string
      change_time:
        type: string
      confirm_pwd:
        example: "111111"
        type: string
      contact:
        type: string
      email:
        type: string
      id:
        example: 5
        type: integer
      ll_time:
        type: string
      nick_name:
        type: string
      phone:
        type: string
      pid:
        type: integer
      privilege_id:
        type: integer
      pwd:
        type: string
      remark:
        type: string
      role_id:
        type: integer
      state:
        type: string
      username:
        type: string
    type: object
...
host: '{{.Host}}'
info:
  contact: {}
  description: jimi co ptt platform web api.
  license: {}
  title: JiMi Web API
  version: 1.0.1
paths:
  /account:
    post:
      parameters:
      - description: 创建下级账户所用的信息
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/model.CreateAccount'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: '{"success": true,"account_id": 1430 }'
          schema:
            type: string
      summary: 创建下级账户
  /account/{account_name}:
    get:
      parameters:
      - description: 当前用户的账号
        in: path
        name: account_name
        required: true
        type: string
      - description: 登录时返回的sessionId
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: "{\"message\":\"User information obtained successfully\",\t\"account_info\":
            ai,\t\"device_list\" deviceAll,\"group_list\": gList,\"tree_data\":resp}"
          schema:
            type: string
      summary: 获取账户信息
  /account/info/update:
    post:
      parameters:
      - description: 登录时返回的sessionId
        in: header
        name: Authorization
        required: true
        type: string
      - description: 更新下级账户的信息，其中id和loginId和NickName不能为空
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/model.AccountUpdate'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: '{"success": "true","msg":"update account success"}'
          schema:
            type: string
      summary: 更新下级账户信息
...
  /account/pwd/update:
    post:
      parameters:
      - description: 登录时返回的sessionId
        in: header
        name: Authorization
        required: true
        type: string
      - description: 更新用户密码
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/model.AccountPwd'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: '{"result": "success","msg":"Password changed successfully"}'
          schema:
            type: string
      summary: 更新用户密码
  /account_class/{accountId}/{searchId}:
    get:
      parameters:
      - description: 当前用户的账号Id
        in: path
        name: accountId
        required: true
        type: string
      - description: 获取用户下级目录的账号Id
        in: path
        name: searchId
        required: true
        type: string
      - description: 登录时返回的sessionId
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: "{\"message\":\"User information obtained successfully\",\t\"account_info\":
            ai,\t\"device_list\" deviceAll,\"group_list\": gList,\"tree_data\":resp}"
          schema:
            type: string
      summary: 获取账户下级目录
...
  /account_device/{accountId}/{getAdviceId}:
    get:
      parameters:
      - description: 当前用户的账号Id
        in: path
        name: accountId
        required: true
        type: string
      - description: 获取用户设备信息的账号Id
        in: path
        name: getAdviceId
        required: true
        type: string
      - description: 登录时返回的sessionId
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: "{\"message\":\"User information obtained successfully\",\t\"account_info\":
            ai,\t\"device_list\" deviceAll,\"group_list\": gList,\"tree_data\":resp}"
          schema:
            type: string
      summary: 获取账户的设备信息
...
swagger: "2.0"